requirements.txt
# requirements.txt
pillow
requests
